requirements.txt
scikit-learn==1.4.0
pandas==2.0.3
numpy==1.26.0
joblib==1.3.2
ipykernel==6.29.0
PyQt6==6.6.1
streamlit==1.28.0
Pillow==10.0.1